Defined new miner interface
