--- internal/domria/image.go
@@ -1,7 +1,0 @@
-package domria
-
-type image struct {
-	flatID int
-	url    string
-	kind   Kind
-}
