--- internal/domria/origin.go
@@ -1,11 +1,0 @@
-package domria
-
-import (
-	"time"
-)
-
-type origin struct {
-	id         int
-	upsertTime time.Time
-	isFound    bool
-}
