--- internal/domria/panorama.go
@@ -1,5 +1,0 @@
-package domria
-
-type panorama struct {
-	Img string `json:"img"`
-}
