--- internal/domria/photo.go
@@ -1,5 +1,0 @@
-package domria
-
-type photo struct {
-	File string `json:"file"`
-}
