--- internal/domria/position.go
@@ -1,6 +1,0 @@
-package domria
-
-type position struct {
-	Lon coordinate `json:"lon"`
-	Lat coordinate `json:"lat"`
-}
