--- internal/domria/prices.go
@@ -1,7 +1,0 @@
-package domria
-
-type prices struct {
-	USD price `json:"1"`
-	EUR price `json:"2"`
-	UAH price `json:"3"`
-}
