--- internal/domria/search.go
@@ -1,5 +1,0 @@
-package domria
-
-type search struct {
-	Items []item `json:"items"`
-}
