+++ internal/mining/miner.go
@@ -1,0 +1,7 @@
+package mining
+
+type Miner interface {
+	Name() string
+	Spec() string
+	MineFlat() (Flat, error)
+}
